backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
chromadb==0.5.0
sentence-transformers==3.0.0
PyMuPDF==1.24.4
python-docx==1.1.2
openpyxl==3.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-slugify==8.0.4
tiktoken==0.7.0
langchain==0.2.1
langchain-community==0.2.1
